wip: move commands into subdirectory
