+++ source/commands/index.ts
@@ -1,0 +1,3 @@
+export {init} from './init.js';
+export {tidy} from './tidy.js';
+export {watch} from './watch.js';
