+++ source/commands/init.ts
@@ -2,9 +2,9 @@ import fs from 'node:fs';
 import path from 'node:path';
 import process from 'node:process';
 import chalk from 'chalk';
-import {defaultConfig, resolveConfig} from './config.js';
-import {writePrettyFile} from './util/write-pretty-file.js';
-import type {UserOptions} from './types/config.js';
+import {defaultConfig, resolveConfig} from '../config.js';
+import {writePrettyFile} from '../util/write-pretty-file.js';
+import type {UserOptions} from '../types/config.js';
 
 export async function init(flags?: UserOptions) {
 	const {options} = await resolveConfig(flags);
